// typechain-types/MessageMetadata.ts
export interface MessageMetadataInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "acknowledgeMessage"
      | "getMessageInfo"
      | "sendMessageCommitment"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "MessageAcknowledged" | "MessageSent"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "acknowledgeMessage",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getMessageInfo",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "sendMessageCommitment",
    values: [BytesLike, string, string]
  ): string;

  decodeFunctionResult(
    functionFragment: "acknowledgeMessage",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getMessageInfo",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "sendMessageCommitment",
    data: BytesLike
  ): Result;
}

export namespace MessageAcknowledgedEvent {
  export type InputTuple = [
    messageHash: BytesLike,
    receiverDID: string,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    messageHash: string,
    receiverDID: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    messageHash: string;
    receiverDID: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace MessageSentEvent {
  export type InputTuple = [
    messageHash: BytesLike,
    senderDID: string,
    receiverDID: string,
    ipfsCid: string,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    messageHash: string,
    senderDID: string,
    receiverDID: string,
    ipfsCid: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    messageHash: string;
    senderDID: string;
    receiverDID: string;
    ipfsCid: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface MessageMetadata extends BaseContract {
  connect(runner?: ContractRunner | null): MessageMetadata;
  waitForDeployment(): Promise<this>;

  interface: MessageMetadataInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  acknowledgeMessage: TypedContractMethod<
    [messageHash: BytesLike],
    [void],
    "nonpayable"
  >;

  getMessageInfo: TypedContractMethod<
    [messageHash: BytesLike],
    [
      [string, string, bigint, boolean, boolean] & {
        senderDID: string;
        receiverDID: string;
        timestamp: bigint;
        acknowledged: boolean;
        exists: boolean;
      }
    ],
    "view"
  >;

  sendMessageCommitment: TypedContractMethod<
    [messageHash: BytesLike, receiverDID: string, ipfsCid: string],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "acknowledgeMessage"
  ): TypedContractMethod<[messageHash: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "getMessageInfo"
  ): TypedContractMethod<
    [messageHash: BytesLike],
    [
      [string, string, bigint, boolean, boolean] & {
        senderDID: string;
        receiverDID: string;
        timestamp: bigint;
        acknowledged: boolean;
        exists: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "sendMessageCommitment"
  ): TypedContractMethod<
    [messageHash: BytesLike, receiverDID: string, ipfsCid: string],
    [void],
    "nonpayable"
  >;

  getEvent(
    key: "MessageAcknowledged"
  ): TypedContractEvent<
    MessageAcknowledgedEvent.InputTuple,
    MessageAcknowledgedEvent.OutputTuple,
    MessageAcknowledgedEvent.OutputObject
  >;
  getEvent(
    key: "MessageSent"
  ): TypedContractEvent<
    MessageSentEvent.InputTuple,
    MessageSentEvent.OutputTuple,
    MessageSentEvent.OutputObject
  >;

  filters: {
    "MessageAcknowledged(bytes32,string,uint256)": TypedContractEvent<
      MessageAcknowledgedEvent.InputTuple,
      MessageAcknowledgedEvent.OutputTuple,
      MessageAcknowledgedEvent.OutputObject
    >;
    MessageAcknowledged: TypedContractEvent<
      MessageAcknowledgedEvent.InputTuple,
      MessageAcknowledgedEvent.OutputTuple,
      MessageAcknowledgedEvent.OutputObject
    >;

    "MessageSent(bytes32,string,string,string,uint256)": TypedContractEvent<
      MessageSentEvent.InputTuple,
      MessageSentEvent.OutputTuple,
      MessageSentEvent.OutputObject
    >;
    MessageSent: TypedContractEvent<
      MessageSentEvent.InputTuple,
      MessageSentEvent.OutputTuple,
      MessageSentEvent.OutputObject
    >;
  };
}